refactoring hasCustomProperties
