+++ vec113/src/main/resources/vec113/vec_1.1.3-ext.xjb
@@ -50,7 +50,11 @@
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='CustomProperty' and @type='vec:CustomProperty']]">
-        <inheritance:implements>com.foursoft.vecmodel.vec113.HasCustomProperties</inheritance:implements>
+        <inheritance:implements>com.foursoft.vecmodel.common.HasCustomProperties&lt;com.foursoft.vecmodel.vec113.VecCustomProperty&gt;</inheritance:implements>
+    </jxb:bindings>
+
+    <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='PropertyType']]">
+        <inheritance:implements>com.foursoft.vecmodel.common.HasPropertyType</inheritance:implements>
     </jxb:bindings>
 
 </jxb:bindings>
