+++ vec120/src/main/resources/vec122/vec_1.2.2-ext.xjb
@@ -50,7 +50,11 @@
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='CustomProperty' and @type='vec:CustomProperty']]">
-        <inheritance:implements>com.foursoft.vecmodel.vec120.HasCustomProperties</inheritance:implements>
+        <inheritance:implements>com.foursoft.vecmodel.common.HasCustomProperties&lt;com.foursoft.vecmodel.vec120.VecCustomProperty&gt;</inheritance:implements>
     </jxb:bindings>
+
+   <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='PropertyType']]">
+       <inheritance:implements>com.foursoft.vecmodel.common.HasPropertyType</inheritance:implements>
+   </jxb:bindings>
 
 </jxb:bindings>
